requirements.txt
Django>=5.0
Pillow>=10.0
qrcode[pil]>=7.0
python-dotenv>=1.0

# Production server
gunicorn>=21.0

# Database
dj-database-url>=2.0
psycopg2-binary>=2.9

# Static files
whitenoise>=6.6

# Cloud storage
cloudinary>=1.37
django-cloudinary-storage>=0.3
